+++ lib/sequel_rails.rb
@@ -1,5 +1,6 @@
 require 'sequel_rails/version'
 require 'sequel_rails/railtie' if defined? Rails
+require 'english'
 
 module SequelRails
   def self.jruby?